Automatically generate full Singularity HTTP client library from swagger
Add a new package under `client`, titled `swagger` and configure
`go:generate` directives to automatically generate a full HTTP client
for singularity in it. This will remove the need for manual
implementation of the go HTTP client library for Singularity,
considering the breadth of its APIs.

The code generation requires `swagger` binary in path. Adjust CI build
to install the required binary and check consistency of the generated
code with the checked-in code as part of the regular build.