Prepare Files For Packing APIs (#192)
# Goals

Provide external apis that will prepare a file or whole datasource for
packing. This also enabled extracting all of the code for the scan step
of the dataset worker.

# Implementation

Create an implentation that, given an fileID, looks up that File's
FileRanges and attempts to put them in one or more PackJobs. For
complete PackJobs, they are left in status `Ready`. For incomplete
PackJobs, they're left `Created` for later, and FileRanges are merged
into future preparation operations.

Providing an incremental "prepare PackJobs by file" API simplifies the
logic of scanning and preparing whole sources for packing. Now we just
incrementally prepare each file in a data source (either with or without
scanning) and at end, we mark any outstanding PackJobs that are
incomplete `Ready`. (this means that if all files in a source are
already prepared for packing, then "preparing a data source" is only
this last call of marking any incomplete pack jobs ready)