Xinaxu/refactor (#221)
This is a big refactor change that combines multiple important and
breaking changes. Almost all files have changed. Hence, the best way to
review this PR is to read all files from this branch instead of relying
on diff view
- [x] datasource and dataset model are now storage and preparation. A
storage is connection to a local or remote storage. A preparation is a
data prep flow that allows multiple source storages and multiple output
storages. There is API to attach source and output storages.
- [x] Age encryption is removed so now one pack job maps to one Car
- [x] All tests have been refactored to test against all 3 SQL
backend(SQLite, MySQL, postgres)
  - Use testutil.All to test with all backend
  - Use testutil.One to test the first available backend
- [x] CLI output are now rendered using
https://github.com/data-preservation-programs/table
- [x] CLI output in tests are now captured and will be output to
cmd/testdata, though it won't be part of pull request because the
tempdir difference
- [x] core pack and daggen code is refactored to implement io.Reader so
it can be directly used with output storage backed by Rclone
- [x] motion related APIs are refactored to provide interface as similar
as possible. This includes
  - [x] table name change in the query and API signature change
  - [x] common code moved from handler package to a separate package
  - [x] GetFileDeals includes #257
  - [x] GetFile
  - [x] PushFile
- [x] GetSourcePackJobs not used. can be replaced by
dataprep.GetStatusHandler
  - [x] PrepareToPackFile 
  - [x] PrepareToPackSource renamed to PrepareToPackRemaining
  - [x] Pack not used
  - [x] Fixes
    - [x] Fix: PackHandler did not mark packJob complete
  - [x] Unit tests for above
- [x] API Integration tests for above (create prep, push file, pack and
get the list)
  - Please take a look at
    - `cmd/api_test.go`
    - `handler/file`
    - `handler/job/pack.go`
- [x] go generate now requires npm to generate typescript client based
on swagger.json
- [x] go generate