Improvements on Makefile (#360)
Some improvements on `Makefile`:

* Dependency check: All targets that require `go` will check if `go` is
installed. If not, it will exit and inform the user to install `go`
instead of giving an error like it does now.
* Dependency check: `lint` will also check if the dependencies it needs
(`golancici-lint` and `staticcheck`) are installed, and if not it will
install them.
* Dependency check: `test` will also check if the dependency it needs
(`gotestsum`) is installed and install it if it isn't.
* Added a `make help` option to inform the user of different targets.
Note that this is now the default behavior so `make` will trigger `make
help` directly, which is usually the recommend approach. If however,
maintaining `make build` as the default is preferred, let me know, as it
can be easily changed back.
* Also note that `install-lint-deps` and `install-test-deps` tagerts are
not included in the help information displayed to the user, as it's not
something they would/should use directly.