Use lotus head latency in deal tracker (#383)
resovles #336 

Previously, we are making a 24 hour buffer time to account for potential
latency of StateMarketDeals call. However, that leads to some deal state
to change between expired and published during 24 hour window as the
buffer time for marking a deal expired comes from the last updated time
inside StateMarketDeal payload.

This change fixes most of the problem by first calling Lotus ChainHead
to determine the current lotus latency first. Edges cases are handled
below
* If the StateMarketDeals response is based on an earlier epoch than
ChainHead response, then it's likely some deals are in expired state
based on ChainHead latency but is not expired in StateMarketDeals
response, the state of the deal will be updated in the next run when
StateMarketDeals response is based on a later epoch
* If the StateMarketDeals response is based on a latter epoch than
ChainHead response, then those deals that are in non_expired state based
on ChainHead latency will be marked as expired