Fix scna concurrency limiting to limit goroutines created
While scan used a worker pool to limit the number of running goroutines, it did not limit the number of goroutines created. The worker pool can be used to do both, and it is not necessary to create a new goroutine that then has the worker pool run a function.