Fix duplicate scan and pack with concurrency greater than 1
When concurrency is n, then n dataset workers are created. They each call scan. Is is necessary to ensure that each dataset worker's scan does not do the same scan as any other dataset worker. This is fixed by setting the transaction isolation level in the finding jobs and marking them as belonging to a worker.

Fixes #388