Optimize retrieval from Filecoin (#404)
Optimize retrieval so that when requested retrieval ranges do not align
with singularity file ranges, only the minimal number of retrieval
requests are made.

This is accomplished by creating a separate reader for each singularity
file range. For reads that are larger than a range, multiple ranges are
read until the read request is satisfied or until all data is read. For
reads smaller than the amount of data remaining in the range, the range
reader is maintained so that it can continue to be read from by
subsequent reads.

This approach associates a reader with each Singularity file range, and
not the ranges requested via the API (in HTTP range header). This avoids
needing to parse the range header in order to create readers where each
reads some number of Singularity ranges. Rather, as arbitrary requested
ranges are read, an existing reader for the corresponding singularity
range(s) is reused if the requested range falls on a singularity range
from a previous read. This also means that there is only a single
retrieval for each singularity range, whereas if readers were associated
with requested ranges then multiple readers could overlap the same
singularity range and require multiple retrievals of the same range.

Fixes
https://github.com/data-preservation-programs/singularity/issues/366
Fixes https://github.com/filecoin-project/motion/issues/143

As an optimization, only one singularity range reader is maintained at a
time. This works because once a new singularity range is selected by the
requested range read, then it is highly unlikely that a subsequent read
request will fall on a
a singularity range that was already read from, previous to the new one.

Additional changes:

- The `filecoinReader` implementation supports the `io.WriteTo`
interface to allow direct copying to an `io.Writer`.
- The `FilecoinRetriever` interface supports the `RetrieveReader`
function that returns an `io.ReadCloser` to read data from.